Rename selector component
